+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - AllenAI s2orc-doc2json as a clowder extractor. [#3](https://github.com/clowder-framework/text-extractor/issues/3)
 - Conversion of json output to text. [#8](https://github.com/clowder-framework/text-extractor/issues/8)
+- Added section headers for text output. [#13](https://github.com/clowder-framework/text-extractor/issues/13)
 
 ### Changed